Update arc-runner.yaml configuration
The arc-runner.yaml configuration of the ArgoCD application has been updated. This change includes an update to the targetRevision and the addition of a service account for the runner controller under the helm values configuration. The service account specifies the namespace and name to be used by the runner controller.