Feat/integrate dev applications (#6)
* feat: add cardio-api

* Add infrastructure application set and arc-systems app

This commit introduces an ArgoCD ApplicationSet for infrastructure applications to automatically sync changes from the specified Git repository. Additionally, a new Application for arc-systems is also included. Changes to the .gitignore file were made to exclude certain YAML files from tracking.

* Update source chart and repoURL in arc.yaml

The source chart and repoURL for 'arc.yaml' in the ArgoCD applications have been updated. The previous chart and repoURL have been commented out, and a new repoURL for the actions-runner-controller-charts/gha-runner-scale-set has been added.

* Update chart and repoURL in arc.yaml

The chart and repoURL under the source in the arc-systems/arc.yaml file have been updated. This change involves switching from an old direct chart URL to a new chart name, while the repoURL has also been updated accordingly.

* Update repoURL in arc.yaml

The repoURL in the arc.yaml file under the infrastructure project has been updated. The URL earlier pointed to an OCI link which has now been changed to a direct link to the GitHub Container Registry.

* Update chart name in arc.yaml

The chart name for gha-runner-scale-set has been updated to gha-runner-scale-set-controller in the infrastructure project of Argo CD. This change correctly aligns the naming with its source repository.

* Update targetRevision in arc.yaml

The targetRevision of the gha-runner-scale-set-controller in arc.yaml has been updated from 0.8.2 to 0.9.1. This modification ensures that we're using the latest stable version of the controller for enhanced reliability and performance.

* Downgrade targetRevision in arc.yaml

The targetRevision for the gha-runner-scale-set-controller in the arc.yaml file has been downgraded. This change is from version 0.9.1 to 0.8.2, potentially due to compatibility issues or to revert to a more stable version.

* Add new arc-crds application and update arc application

A new Argo CD application named arc-crds has been added to handle the deployment of Custom Resource Definitions (CRDs) for the ARC Systems infrastructure. This includes enabling the recursive browsing of the directory, among other specifications. Additionally, in the arc.yaml file, a new helm key has been added to the source section, effectively enabling the skipping of CRDs during the deployment process.

* Add ArgoCD arc-runner application and related resources

This commit adds the arc-runner app and the associated files, including a YAML definition for the arc-runner, a docker daemon config for Kubernetes, and a sealed secret for secure data handling. All these resources aid in the handling of GitHub actions runner infrastructure.

* Update arc-runner.yaml configuration

The arc-runner.yaml configuration of the ArgoCD application has been updated. This change includes an update to the targetRevision and the addition of a service account for the runner controller under the helm values configuration. The service account specifies the namespace and name to be used by the runner controller.

* Update arc-runner.yaml configuration

The arc-runner.yaml configuration of the ArgoCD application has been updated. This change includes an update to the targetRevision and the addition of a service account for the runner controller under the helm values configuration. The service account specifies the namespace and name to be used by the runner controller.

* update runner

* update runner

* add cert-manager

* split issuers

* add nginx

* add nginx

* add nginx

* add nginx

* add nfs-server-provisioner

* add oauth2-proxy to cardio project

* add oauth2-proxy to cardio project

* add oauth2-proxy to cardio project

* add oauth2-proxy to cardio project

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

* add airflow

---------

Co-authored-by: Johannes Otto <jotto@tcc-clinicalsolutions.de>