Disable Application and Helm CRDs in Argo CD
In the Argo CD application configuration for the 'arc-systems' infrastructure, the entire Application object definition in the 'arc-crds.yaml' file was commented out. Similarly, the 'skipCrds: true' Helm option under the source in 'arc.yaml' was also commented out. These changes effectively disable the use of the Application and the skipping of Helm's Custom Resource Definitions (CRDs) within this particular Argo CD setup.