Add hl7-receiver-dev deployment and service
This commit introduces the new "hl7-receiver-dev" Deployment, Service, and SealedSecret in the telehealth-data-integration namespace. The Deployment uses a telehealth-data-integration image and contains environment variables from a secret. Sealed secrets are used to store these environment variables securely. The Service is type of ClusterIP and it exposes port 8000 for TCP traffic.