Add Telehealth Actions Runner Docker configurations
This commit introduces Docker configuration for the Telehealth Actions Runner. It includes the creation of a Dockerfile and a docker-compose.yml file for setting up the Actions Runner as well as the runner's specific configurations within the values.telehealth.yaml file. With these configurations, the runner can be built and deployed within the Telehealth context.